fixed the bad flagtest (and moved it from playerTests to flagtests), fixed a bug in player.pickupflag and added a new testboard for flagtests
