+++ src/main/java/inf112/skeleton/app/Game.java
@@ -109,7 +109,7 @@ public class Game implements ApplicationListener {
         camera = new OrthographicCamera(Gdx.graphics.getWidth(), Gdx.graphics.getHeight());
 
         board = new Board();            //Initialize a board
-        board.readBoard(2);  //Read board info from file (for now hardcoded to 1 since we only have Board1.txt)
+        board.readBoard(120);  //Read board info from file (for now hardcoded to 1 since we only have Board1.txt)
 
         spriteMap = new HashMap<>();
         //For all game objects on map, add the identifying string and a corresponding sprite to spriteMap.
