+++ src/main/java/inf112/skeleton/app/sprites/Player.java
@@ -95,7 +95,7 @@ public class Player extends AbstractGameObject {
     public void pickupFlag() {
         if (board.getOriginalPosition(getCoordinates().getX(), getCoordinates().getY()).getName().matches("Flag\\d+")) {
             System.out.println("Player on flag");
-            Flag flag = (Flag) board.getPosition(getCoordinates().getX(), getCoordinates().getY());
+            Flag flag = (Flag) board.getOriginalPosition(getCoordinates().getX(), getCoordinates().getY());
             if (flag.pickUp(this)) {
                 addScore(1);
                 visitedFlags.add(flag.getShortName());
