+++ src/main/resources/boards/Board120.txt
@@ -1,0 +1,12 @@
+g h g g g g g g g g g g
+g g g g g g g g g w f3 g
+g g g g w w g g g w h g
+g g g f2 g g g g g w w g
+g g p1 f1 g g g g g g g g
+g g g g g g g p3 g g g g
+g g g g g g g g g g g g
+g g g g w w w w g g g g
+g g g g g g g w g g g g
+g g g g g g g w g g g g
+g g g p2 g g g w g g g g
+g g g g g g g g g g p4 g
